+++ include/wave/geometry/src/debug/PrintExpression.hpp
@@ -7,28 +7,10 @@
 
 #include <iostream>
 #include <regex>
-
-#include <boost/version.hpp>
-#if BOOST_VERSION < 105600
-#include <boost/units/detail/utility.hpp>
-#else
 #include <boost/core/demangle.hpp>
-#endif
 
 namespace wave {
 namespace internal {
-
-/** Gets demangled name of a type_info object - call as demangleName(typeid(T));
- *
- * Wrapper to handle different versions of Boost
- */
-inline std::string demangleName(const std::type_info &ti) {
-#if BOOST_VERSION < 105600
-    return boost::units::detail::demangle(ti.name());
-#else
-    return boost::core::demangle(ti.name());
-#endif
-}
 
 /** Get the unqualified template name of a type name */
 inline std::string getTemplateName(const std::string &full_name, bool include_tp) {
@@ -51,7 +33,7 @@ inline std::string getTemplateName(const std::string &full_name, bool include_tp
 template <typename T>
 std::string getTypeString() {
     std::stringstream s;
-    const auto &name = demangleName(typeid(T));
+    const auto &name = boost::core::demangle((typeid(T).name()));
     s << (std::is_const<T>{} ? "const" : "");
     s << getTemplateName(name, true);
     s << (std::is_lvalue_reference<T>{} ? "&" : "");
@@ -62,7 +44,7 @@ std::string getTypeString() {
 /** Get the unqualified typename of an expression for debugging purposes.*/
 template <typename Derived>
 inline std::string getExpressionTypeName(const ExpressionBase<Derived> &) {
-    const auto full_name = demangleName(typeid(Derived));
+    const auto full_name = boost::core::demangle((typeid(Derived).name()));
 
     // For leaf expression, include the template parameters (typically ImplType)
     const bool include_template_parameters = is_leaf_expression<Derived>();
@@ -75,7 +57,7 @@ inline std::string getExpressionTypeName(const ExpressionBase<Derived> &) {
  * This variation includes the first template parameter (the To type) */
 template <typename To, typename From>
 inline std::string getExpressionTypeName(const Conversion<To, From> &) {
-    const auto to_name = demangleName(typeid(To));
+    const auto to_name = boost::core::demangle((typeid(To).name()));
     return "Conversion to " + getTemplateName(to_name, true);
 }